agents/claude-research/pyproject.toml
[project]
name = "claude-research"
version = "2.0.1"
description = "Multi-agent research system with PDF/PPTX output capabilities"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk",
    "reportlab>=4.0.0",
    "python-pptx>=0.6.21",
    "matplotlib>=3.7.0",
    "pillow>=10.0.0",
    "shared",
]

[tool.uv.sources]
shared = { path = "../../shared" }

[project.optional-dependencies]
dev = [
    "python-dotenv",
]

[project.scripts]
run-agent = "research_agent:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["research_agent*"]

[tool.setuptools.package-data]
"research_agent.prompts" = ["*.txt"]
